Added Costum BRDF panel.
You can add a new costum brdf as you prefer. You need to give it a name and click on the + button. This will add it to the temporary (Costum BRDFs dropdown). If you want to put it in the permenant BRDFs, you need at least to have a working BRDF then you press ADD to add it to the permenant BRDFs.