Add save brdf button
Save BRDF will save the BRDF to the local default BRDF folder. You can not delete the BRDFs once they are created, noor change their names. If you want to change the BRDF name or delete it, you will need to do that manually. This is a safety feature implemented to the engine.